+++ src/tui/action.rs
@@ -79,6 +79,7 @@ make_actions! {
     RemoveLine "Remove Line" key!(d),
     SaveDrawer "*S*ave Drawer" key!(ctrl-S),
     Search "Search" key!('/'),
+    Sort "Sort",
     ToggleHiding "Toggle *H*iding" key!(ctrl-H), // hiding either pwd chars or unselected values
     ToggleMarkdown "Toggle Markdown",
 }